Move detect_sqlcipher_version to db/misc.py
This way we avoid importing anything that logs at trace. So the sanity
check of packaging (also in Dockerfile) does not hit an attribute
error for log.trace.
