+++ rotkehlchen/db/misc.py
@@ -1,0 +1,22 @@
+"""A module with miscellaneous functions that do not depend on db/drivers/gevent.py
+due. Needed in packaging since that code has log.trace and should only be
+imported if trace level has been setup"""
+
+import re
+
+from pysqlcipher3 import dbapi2 as sqlcipher
+
+
+def detect_sqlcipher_version() -> int:
+    """Returns the major part of the version of the system's sqlcipher package"""
+    conn = sqlcipher.connect(':memory:')  # pylint: disable=no-member
+    query = conn.execute('PRAGMA cipher_version;')
+    version = query.fetchall()[0][0]
+
+    match = re.search(r'(\d+).(\d+).(\d+)', version)
+    if not match:
+        raise ValueError(f'Could not process the version returned by sqlcipher: {version}')
+
+    sqlcipher_version = int(match.group(1))
+    conn.close()
+    return sqlcipher_version
